Fix array value for attribute handling in source generator and add .gitignore rules
This commit introduces modifications to the handler of attributes in SourceGenerators.Property, specifically when dealing with arrays. A new extension method ToCSharp was added to TypedConstant class which is used to handle attribute constructor arguments. In addition, rules were added to the .gitignore file to exclude GitHub Copilot chat sessions. This commit also includes a refinement in the use of using-strings in the PropertyIncrementalSourceGenerator class, using manual addition to the set instead of set.AddRange() method.