Skip processing fields with "<" as first character
Additional condition added in PropertyIncrementalSourceGenerator to ignore any field whose name starts with "<". This change is necessary to avoid generation for backing fields where this character is common but does not affect other fields (as it is not a valid name for a field). Accompanying unit test "PropertiesBackingFieldIsLeftUnaffected" was also added to ensure the effectiveness of the fix.