Add documentation takeover feature to property generator
This update introduces a feature to the property generator that takes the documentation from the fields and copies it over to the properties. The change includes additions to the source generator code and corresponding updates to the tests. This enhancement also simplifies the workflow as there is no need to separately document the generated properties. Version also was bumped to 2.1.0.