Remove redundant parameter in attribute usage examples
Updated the README to show that the NotifyPropertyChanged and NotifyPropertyChanging attributes default to true, thus eliminating the need for explicitly passing true as a parameter. Also, the default was written as being false but is actually true, making the documentation missmatch the reality. Both have been adressed in this commit.