Move attributes into their own project and separate generator and attributes
This commit concludes a change that was triggered by increasing warnings once the package was used by two or more assemblies, namely: The always re-generated attributes caused a collission warning, that bloats the code output report. This commit fixes that by introducing a separate package for the abstraction, which replaces the existing package (kind of making it the other way around ... so technically, i guess, the "new package" is the source generator? but practically, i did not create that newly but only renamed it... you get it i think).

As this might break someones use, i decided to bump the major version in accordance with semver. It technically does not require any migration what-so-ever, unless you explicitly loaded the corresponding types from your very assembly.